Creation of the Sign Up and Log In interface and data validation
The Sign Up and Log In user interface was created, adding connection between both modules, as well as validating user data in both modules so that an account cannot be created if the data entered does not meet the requirements. and that you cannot log in if the field is empty, leaving open the implementation of account validation within the database.